# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Simple Inventory API",
    "description": "This is a simple API",
    "contact": {
      "email": "you@your-company.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/VITYLJATROJAN/nika-gold/2.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
...
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Отримати список всіх користувачів",
        "responses": {
          "200": {
            "description": "Список користувачів",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          },
          "403": { "description": "Доступ заборонений" },
          "500": { "description": "Внутрішня помилка сервера" }
        }
      }
    },
    "/api/admin/users/{id}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "Видалити користувача",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ідентифікатор користувача",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Користувача видалено" },
          "404": { "description": "Користувач не знайдений" },
          "500": { "description": "Внутрішня помилка сервера" }
        }
      }
    },
...
    "/api/admin/finance/online/orders": {
      "get": {
        "tags": ["Finance - Online"],
        "summary": "Отримати всі онлайн-замовлення",
        "description": "Повертає список усіх онлайн-замовлень, включаючи їхній статус, продукти, загальну суму і методи оплати.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список онлайн-замовлень успішно отримано.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "ID замовлення"
                      },
                      "products": {
                        "type": "array",
                        "description": "Список продуктів у замовленні.",
                        "items": {
                          "type": "object",
                          "properties": {
                            "productId": {
                              "type": "string",
                              "description": "ID продукту"
                            },
                            "quantity": {
                              "type": "integer",
                              "description": "Кількість"
                            },
                            "name": {
                              "type": "string",
                              "description": "Назва продукту"
                            }
                          }
                        }
                      },
                      "totalPrice": {
                        "type": "number",
                        "description": "Загальна сума замовлення"
                      },
                      "paymentMethod": {
                        "type": "string",
                        "description": "Метод оплати"
                      },
                      "status": {
                        "type": "string",
                        "description": "Статус замовлення"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Токен відсутній або недійсний." },
          "403": {
            "description": "Доступ заборонено для неадміністративних користувачів."
          }
        }
      },
      "post": {
        "tags": ["Finance - Online"],
        "summary": "Створити нове онлайн-замовлення",
        "description": "Додає нове онлайн-замовлення з вказаними продуктами, сумою і методом оплати.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "products": {
                    "type": "array",
                    "description": "Продукти у замовленні.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "ID продукту"
                        },
                        "quantity": {
                          "type": "integer",
                          "description": "Кількість продукту"
                        }
                      }
                    }
                  },
                  "totalPrice": {
                    "type": "number",
                    "description": "Загальна сума замовлення"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "description": "Метод оплати"
                  },
                  "userId": {
                    "type": "string",
                    "description": "ID користувача"
                  }
                },
                "example": {
                  "products": [
                    { "productId": "12345", "quantity": 2 },
                    { "productId": "67890", "quantity": 1 }
                  ],
                  "totalPrice": 500,
                  "paymentMethod": "card",
                  "userId": "user123"
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Онлайн-замовлення успішно створено." },
          "400": { "description": "Недійсні дані у запиті." }
        }
      }
    },
    "/api/admin/finance/online/orders/{id}": {
      "get": {
        "tags": ["Finance - Online"],
        "summary": "Отримати конкретне онлайн-замовлення",
        "description": "Повертає детальну інформацію про конкретне онлайн-замовлення за ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID онлайн-замовлення.",
            "schema": { "type": "string", "example": "order123" }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "responses": {
          "200": { "description": "Деталі замовлення успішно отримано." },
          "404": { "description": "Замовлення не знайдено." }
        }
      },
      "patch": {
        "tags": ["Finance - Online"],
        "summary": "Оновити статус онлайн-замовлення",
        "description": "Оновлює статус конкретного онлайн-замовлення (наприклад, 'completed' або 'cancelled').",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID онлайн-замовлення.",
            "schema": { "type": "string", "example": "order123" }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "description": "Новий статус замовлення.",
                    "enum": ["new", "completed", "cancelled"]
                  }
                },
                "example": { "status": "completed" }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Статус замовлення успішно оновлено." },
          "400": { "description": "Недійсний статус." },
          "404": { "description": "Замовлення не знайдено." }
        }
      },
      "put": {
        "tags": ["Finance - Online"],
        "summary": "Оновити онлайн-замовлення",
        "description": "Оновлює дані конкретного онлайн-замовлення, включаючи адресу доставки, статус та інші параметри.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID онлайн-замовлення.",
            "schema": { "type": "string", "example": "order123" }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": { "type": "string", "example": "Bearer eyJhbGciOiJI..." }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "description": "Новий статус замовлення."
                  },
                  "paymentMethod": {
                    "type": "string",
                    "description": "Оновлений метод оплати."
                  },
                  "deliveryAddress": {
                    "type": "string",
                    "description": "Оновлена адреса доставки."
                  }
                },
                "example": {
                  "status": "assembled",
                  "paymentMethod": "card",
                  "deliveryAddress": "Новий проспект, 25"
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "✅ Замовлення успішно оновлено." },
          "400": { "description": "❌ Недійсні дані." },
          "404": { "description": "❌ Замовлення не знайдено." }
        }
      }
    },
    "/api/admin/finance/online/orders/{id}/return": {
      "put": {
        "tags": ["Finance - Online"],
        "summary": "Часткове повернення товару",
        "description": "Дозволяє користувачеві повернути окремі товари із замовлення.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID онлайн-замовлення.",
            "schema": { "type": "string", "example": "order123" }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": { "type": "string", "example": "Bearer eyJhbGciOiJI..." }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "returnedProducts": {
                    "type": "array",
                    "description": "Список товарів, що повертаються.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "ID продукту."
                        },
                        "quantity": {
                          "type": "integer",
                          "description": "Кількість поверненого товару."
                        }
                      }
                    }
                  },
                  "refundAmount": {
                    "type": "number",
                    "description": "Сума відшкодування."
                  },
                  "updatedBy": {
                    "type": "string",
                    "description": "ID адміністратора, який обробив повернення."
                  }
                },
                "example": {
                  "returnedProducts": [
                    { "productId": "product123", "quantity": 2 }
                  ],
                  "refundAmount": 150,
                  "updatedBy": "admin789"
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "✅ Часткове повернення виконано." },
          "400": { "description": "❌ Некоректні дані." },
          "404": { "description": "❌ Замовлення не знайдено." }
        }
      }
    },
...
    "/api/admin/finance/offline/orders": {
      "get": {
        "tags": ["Finance - Offline"],
        "summary": "Отримати всі офлайн-замовлення",
        "description": "Повертає список усіх офлайн-замовлень із інформацією про продукти, статус, метод оплати і місце замовлення.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список офлайн-замовлень успішно отримано.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "ID замовлення"
                      },
                      "products": {
                        "type": "array",
                        "description": "Список продуктів у замовленні.",
                        "items": {
                          "type": "object",
                          "properties": {
                            "productId": {
                              "type": "string",
                              "description": "ID продукту"
                            },
                            "quantity": {
                              "type": "integer",
                              "description": "Кількість"
                            },
                            "name": {
                              "type": "string",
                              "description": "Назва продукту"
                            }
                          }
                        }
                      },
                      "totalPrice": {
                        "type": "number",
                        "description": "Загальна сума замовлення"
                      },
                      "paymentMethod": {
                        "type": "string",
                        "description": "Метод оплати"
                      },
                      "saleLocation": {
                        "type": "string",
                        "description": "Місце замовлення"
                      },
                      "status": {
                        "type": "string",
                        "description": "Статус замовлення"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Токен відсутній або недійсний." },
          "403": {
            "description": "Доступ заборонено для неадміністративних користувачів."
          }
        }
      },
      "post": {
        "tags": ["Finance - Offline"],
        "summary": "Створити нове офлайн-замовлення",
        "description": "Додає нове офлайн-замовлення з вказаними продуктами, сумою, методом оплати і місцем замовлення.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "products": {
                    "type": "array",
                    "description": "Продукти у замовленні.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "ID продукту"
                        },
                        "quantity": {
                          "type": "integer",
                          "description": "Кількість продукту"
                        },
                        "name": {
                          "type": "string",
                          "description": "Назва продукту"
                        },
                        "price": {
                          "type": "number",
                          "description": "Ціна за одиницю"
                        }
                      }
                    }
                  },
                  "totalPrice": {
                    "type": "number",
                    "description": "Загальна сума замовлення"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "description": "Метод оплати"
                  },
                  "saleLocation": {
                    "type": "string",
                    "description": "Місце замовлення"
                  },
                  "status": {
                    "type": "string",
                    "description": "Статус замовлення",
                    "enum": ["pending", "completed", "cancelled"]
                  }
                },
                "example": {
                  "products": [
                    {
                      "productId": "12345",
                      "quantity": 2,
                      "name": "Product Name",
                      "price": 250
                    },
                    {
                      "productId": "67890",
                      "quantity": 1,
                      "name": "Another Product",
                      "price": 300
                    }
                  ],
                  "totalPrice": 800,
                  "paymentMethod": "cash",
                  "saleLocation": "Shop #45",
                  "status": "pending"
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Офлайн-замовлення успішно створено." },
          "400": { "description": "Недійсні дані у запиті." }
        }
      }
    },
    "/api/admin/finance/offline/orders/{id}": {
      "patch": {
        "tags": ["Finance - Offline"],
        "summary": "Оновити статус офлайн-замовлення",
        "description": "Оновлює статус конкретного офлайн-замовлення за ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID офлайн-замовлення.",
            "schema": { "type": "string", "example": "order123" }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "description": "Новий статус замовлення.",
                    "enum": ["pending", "completed", "cancelled"]
                  }
                },
                "example": { "status": "completed" }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Статус офлайн-замовлення успішно оновлено."
          },
          "400": { "description": "Недійсний статус." },
          "404": { "description": "Офлайн-замовлення не знайдено." }
        }
      },
      "get": {
        "summary": "Отримати офлайн-замовлення за ID",
        "description": "Повертає деталі офлайн-замовлення та список товарів.",
        "tags": ["Finance - Offline"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID офлайн-замовлення",
            "schema": {
              "type": "string",
              "example": "664b34c2e12f87a5b2d598ac"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успішне отримання замовлення",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "offlineOrder": {
                      "$ref": "#/components/schemas/OfflineOrder"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Замовлення не знайдено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Offline order not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Помилка сервера при отриманні замовлення",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch offline order"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user": {
      "get": {
        "summary": "Get all products",
        "description": "Fetches all available products.",
        "tags": ["User"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successfully fetched the product list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": { "type": "string" },
                      "name": { "type": "string" },
                      "price": { "type": "number" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized - No user ID found" },
          "500": { "description": "Server error" }
        }
      }
    },
...
    "/api/user/orders/{id}/received": {
      "patch": {
        "tags": ["User"],
        "summary": "Підтвердження отримання товару",
        "description": "Оновлює статус замовлення на `completed`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID замовлення для підтвердження.",
            "schema": {
              "type": "string",
              "example": "665acb1fd8477d002a8e2f49"
            }
          }
        ],
        "responses": {
          "200": { "description": "Отримання підтверджено." },
          "500": { "description": "Помилка." }
        }
      }
    },
    "/api/user/orders/purchase-history": {
      "get": {
        "tags": ["User"],
        "summary": "Історія покупок користувача",
        "description": "Повертає список завершених замовлень.",
        "responses": {
          "200": { "description": "Історія покупок отримана." },
          "500": { "description": "Помилка." }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "required": ["email", "password", "username"],
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "654321abcdef" },
          "username": { "type": "string", "description": "Ім'я користувача" },
          "email": {
            "type": "string",
            "description": "Електронна пошта",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Пароль",
            "format": "password"
          },
          "avatar": { "type": "string", "description": "URL аватара" },
          "basket": {
            "type": "array",
            "description": "Кошик користувача",
            "items": { "type": "string" }
          },
          "createdAt": {
            "type": "string",
            "description": "Дата створення аккаунту",
            "format": "date-time"
          },
          "role": {
            "type": "string",
            "description": "Роль ('user', 'admin')",
            "example": "user"
          },
          "tokenStatus": {
            "type": "string",
            "description": "Статус токена (активний, прострочений)"
          },
          "isVerified": {
            "type": "boolean",
            "description": "Чи верифікований email",
            "example": true
          },
          "Wallet": {
            "type": "object",
            "properties": {
              "wallet": {
                "type": "number",
                "description": "Поточний баланс гаманця користувача",
                "example": 42.75
              }
            }
          }
        }
      },
...
    }
  }
}